Cria tabela Vendas
