+++ src/pages/Vendas/Vendas.module.css
@@ -1,0 +1,14 @@
+.container {
+    width: 50em;
+    margin: 0 auto;
+    padding: 40px 70px;
+    background-color: var(--background);
+  }
+  
+  .headerContainer {
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-end;
+    padding: 15px 0;
+  }
+
