+++ src/pages/Vendas/VendasLista/VendasLista.module.css
@@ -1,0 +1,24 @@
+.section{
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    padding: 1rem;
+    align-items: center;
+    align-content: center;
+  };
+  
+.table{
+    margin: 1rem;
+  }
+
+.Container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+a {
+    color: blueviolet;
+    margin-left: 5%;
+    margin-right: 5%;
+}
